feat(fnp-bitmap-facade-1): Spark-exact degrees/radians UDFs with refusal and ANSI STRING cast
New spark_degrees module owns the pair for both doors: user-defined
coercion over numerics and STRING, planning refusal with Spark's class
and text for every other type, the bit-exact single multiply, and
CAST_INVALID_INPUT under ANSI (NULL with ANSI off). The facade drops
its Python double cast; dispatch rebinds to the same UDFs. The root
pays net-negative by inlining its single-use date-shim helper.

Authored-By: Muse Spark (muse-spark-1.3-contributor) <noreply@meta.ai>

+++ python/repark/src/repark/spark/functions_math.py
@@ -154,14 +154,14 @@ def degrees(col: Column | str) -> Column:
     """Radians to degrees (PySpark ``functions.degrees``)."""
     column = _as_column_arg(col, as_lit=False)
     display = f"DEGREES({column.spark_wrap_display_part()})"
-    return _scalar("degrees", column.cast("double"), display=display)
+    return _scalar("degrees", column, display=display)
 
 
 def radians(col: Column | str) -> Column:
     """Degrees to radians (PySpark ``functions.radians``)."""
     column = _as_column_arg(col, as_lit=False)
     display = f"RADIANS({column.spark_wrap_display_part()})"
-    return _scalar("radians", column.cast("double"), display=display)
+    return _scalar("radians", column, display=display)
 
 
 def toDegrees(col: Column | str) -> Column:  # noqa: N802
